refactor the routes and comment the code
